Address comments part-1
